DHFPROD-4519: Add JobsManager test
