+++ one-ui/build.gradle
@@ -65,6 +65,8 @@ dependencies {
 	testRuntime "org.junit.jupiter:junit-jupiter-engine:${junitJupiterVersion}"
 
 	testCompile "org.junit.platform:junit-platform-runner:${junitPlatformVersion}"
+    testCompile group: 'org.mockito', name: 'mockito-core', version: '3.0.0'
+    testCompile group: 'org.mockito', name: 'mockito-junit-jupiter', version: '3.0.0'
 }
 
 task installDependencies(type: NpmTask, group: taskGroup) {
